backend/requirements.txt
fastapi>=0.127.0
uvicorn[standard]>=0.40.0
motor>=3.7.1
pydantic>=2.12.5
pydantic-settings>=2.12.0
python-dotenv>=1.2.1
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.21
google-api-python-client>=2.187.0
google-auth>=2.45.0
google-auth-httplib2>=0.3.0
google-auth-oauthlib>=1.2.2
email-validator>=2.3.0
pyjwt>=2.8.0
cryptography>=41.0.0

